refactor : Separate feature
1. 효과 분리
2. 데시벨 측정 분리
3. 해 위치 분리
4. 해 위치 화면 기준 비율로 수정
