+++ frontend/src/pages/InGame/Mission4/Mission4.js
@@ -2,7 +2,8 @@ import React, { useContext, useRef, useEffect, useState } from 'react';
 import { OpenViduContext, GameContext } from '../../../contexts';
 import styled from 'styled-components';
 import { GameLoading } from '../components';
-import confetti from 'canvas-confetti';
+import { rainEffect, effect } from './effect';
+import { calculateDecibels } from './decibelUtils';
 import sunImage from '../../../assets/sun.png';
 import hillImage from '../../../assets/hill.png';
 
@@ -14,7 +15,6 @@ const Mission4 = () => {
   const [stream, setStream] = useState(null);
   const [decibels, setDecibels] = useState(0); // 데시벨 상태
   const [shoutingDuration, setShoutingDuration] = useState(0); // 함성이 지속된 시간
-  //const [showFailEffect, setShowFailEffect] = useState(false);
   const [sunPositionY, setSunPositionY] = useState(window.innerHeight); // 해의 Y 위치
   const [elapsedTime, setElapsedTime] = useState(0); // 경과 시간 (초 단위)
   const startTimeRef = useRef(null); // 시작 시간 저장
@@ -39,23 +39,24 @@ const Mission4 = () => {
 
     // 매 초마다 경과 시간을 업데이트
     const intervalId = setInterval(() => {
-      const now = Date.now();
-      const elapsedSeconds = Math.floor((now - startTimeRef.current) / 1000);
+      const elapsedSeconds = Math.floor(
+        (Date.now() - startTimeRef.current) / 1000,
+      );
       setElapsedTime(elapsedSeconds);
-      console.log('Elapsed Time: ', elapsedSeconds);
 
       // 시간이 제한 시간보다 많으면 실패 플래그 설정
       if (elapsedSeconds > timeLimit) {
-        // setShowFailEffect(true);
         clearInterval(intervalId);
         setIsGameOver(true);
       }
     }, 1000);
+
+    return () => clearInterval(intervalId);
   }, []);
 
   useEffect(() => {
     if (myMissionStatus && !isGameOver) {
-      effect();
+      effect(3);
     }
   }, [myMissionStatus]);
 
@@ -64,9 +65,10 @@ const Mission4 = () => {
 
     if (elapsedTime > timeLimit && isGameOver) {
       console.log('Challenge failed!');
-      rainEffect(canvasRef);
+      rainEffect(canvasRef, 3);
       return;
     }
+
     const audioContext = new AudioContext();
     const source = audioContext.createMediaStreamSource(stream);
     const analyser = audioContext.createAnalyser();
@@ -76,47 +78,28 @@ const Mission4 = () => {
     const bufferLength = analyser.frequencyBinCount;
     const dataArray = new Uint8Array(bufferLength);
 
-    const id = setInterval(() => {
-      analyser.getByteTimeDomainData(dataArray);
+    const intervalId = setInterval(() => {
+      const decibels = calculateDecibels(analyser, dataArray, bufferLength);
+      setDecibels(decibels); // 데시벨 상태 업데이트
 
-      let sum = 0;
-      for (let i = 0; i < bufferLength; i++) {
-        const value = (dataArray[i] - 128) / 128;
-        sum += value * value;
-      }
-
-      const rms = Math.sqrt(sum / bufferLength);
-      const newDecibels = 40 * Math.log10(rms) + 100;
-      setDecibels(newDecibels); // 데시벨 상태 업데이트
-      console.log(`Decibels: ${newDecibels.toFixed(2)} dB`);
-
-      if (newDecibels > 50) {
+      if (decibels > 50) {
         setShoutingDuration(prevDuration => prevDuration + 0.2);
-      } else {
-        ////// 지속하지 않을 경우 초기화
-        //// setShoutingDuration(0)
       }
+      // else {
+      //   //// 지속하지 않을 경우 초기화
+      //   // setShoutingDuration(0)
+      // }
       if (shoutingDuration > 5) {
-        console.log('Challenge passed!');
-        clearInterval(id);
+        clearInterval(intervalId);
         setMyMissionStatus(true);
         // firework();
         return;
       }
-      console.log('isGameOver: ', isGameOver);
-
-      const maxSunPositionY = 50; // 해가 화면 상단에 위치하는 최소 값
-      const newSunPositionY = Math.max(
-        window.innerHeight - shoutingDuration * 120,
-        maxSunPositionY,
-      );
-      setSunPositionY(newSunPositionY);
-
-      console.log('=======Shouting Duration: ', shoutingDuration);
+      setSunPosition();
     }, 200);
 
     return () => {
-      clearInterval(id);
+      clearInterval(intervalId);
       audioContext.close();
     };
   }, [stream, isGameLoading, shoutingDuration, isGameOver]);
@@ -129,14 +112,27 @@ const Mission4 = () => {
     }
   }
 
-  // Canvas에 데시벨 그리기//
-  useEffect(() => {
-    if (shoutingDuration <= 5) {
-      const canvas = canvasRef.current;
-      const ctx = canvas.getContext('2d');
-      ctx.clearRect(0, 0, canvas.width, canvas.height);
-    }
-  }, [decibels]);
+  // function setSunPosition() {
+  //   const maxSunPositionY = 50; // 해가 화면 상단에 위치하는 최소 값
+  //   const newSunPositionY = Math.max(
+  //     window.innerHeight - shoutingDuration * 120,
+  //     maxSunPositionY,
+  //   );
+  //   setSunPositionY(newSunPositionY);
+  // }
+
+  function setSunPosition() {
+    const screenHeight = window.innerHeight; // 화면 높이
+    const minPercentage = 10; // 해가 화면 상단에 위치하는 최소 퍼센트 값
+    const percentage = Math.max(
+      ((shoutingDuration * 120) / screenHeight) * 150,
+      minPercentage,
+    );
+
+    // 퍼센트를 높이로 변환하여 위치 설정
+    const newSunPositionY = screenHeight * (1 - percentage / 100);
+    setSunPositionY(newSunPositionY);
+  }
 
   return (
     <>
@@ -161,34 +157,6 @@ const Mission4 = () => {
 };
 
 export default Mission4;
-
-const effect = () => {
-  var end = Date.now() + 5 * 1000;
-
-  // go Buckeyes!
-  var colors = ['#F0F3FF', '#15F5BA'];
-
-  (function frame() {
-    confetti({
-      particleCount: 2,
-      angle: 60,
-      spread: 55,
-      origin: { x: 0 },
-      colors: colors,
-    });
-    confetti({
-      particleCount: 2,
-      angle: 120,
-      spread: 55,
-      origin: { x: 1 },
-      colors: colors,
-    });
-
-    if (Date.now() < end) {
-      requestAnimationFrame(frame);
-    }
-  })();
-};
 
 const FullScreenCanvas = styled.div`
   position: absolute;
@@ -244,68 +212,6 @@ const SoundIndicator = styled.div`
   border: 1px solid ${({ theme }) => theme.colors.primary.light};
   transition: width 0.2s ease; // 너비 변화를 0.5초 동안 부드럽게 애니메이션
 `;
-
-const rainEffect = canvasRef => {
-  const canvas = canvasRef.current;
-  const context = canvas.getContext('2d');
-  const drops = [];
-  let isRaining = true;
-
-  class Drop {
-    constructor(x, y, speed, length) {
-      this.x = x;
-      this.y = y;
-      this.speed = speed;
-      this.length = length;
-      this.draw();
-    }
-
-    draw() {
-      context.beginPath();
-      context.strokeStyle = 'rgba(175, 238, 253, 0.7)';
-      context.moveTo(this.x, this.y);
-      context.lineTo(this.x, this.y + this.length);
-      context.stroke();
-      context.closePath();
-    }
-  }
-
-  function render() {
-    if (!isRaining) return;
-
-    context.clearRect(0, 0, canvas.width, canvas.height);
-
-    drops.forEach(drop => {
-      drop.y += drop.speed;
-      if (drop.y > canvas.height) {
-        drop.y = 0;
-        drop.x = Math.random() * canvas.width;
-        drop.speed = Math.random() * 3 + 1;
-        drop.length = Math.random() * 5 + 2;
-      }
-      drop.draw();
-    });
-
-    requestAnimationFrame(render);
-  }
-
-  let tempX, tempY, tempSpeed, tempLength;
-  for (let i = 0; i < 200; i++) {
-    tempX = Math.random() * canvas.width;
-    tempY = Math.random() * canvas.height;
-    tempSpeed = Math.random() * 3 + 1;
-    tempLength = Math.random() * 5 + 2;
-
-    drops.push(new Drop(tempX, tempY, tempSpeed, tempLength));
-  }
-
-  render();
-
-  setTimeout(() => {
-    isRaining = false;
-    context.clearRect(0, 0, canvas.width, canvas.height);
-  }, 3000);
-};
 
 const Sun = styled.div`
   position: absolute;
